Additional Test Cases & Edge Case Robustness (#30)
* Add climits for linux compilation

* Adds a post-processing step to Tabucol which both minimizes the
colors used during the random coloring process and makes them
contiguous. Closes #28

* Adds a check to Tabucol to guarantee that we choose a reasonable
starting condition (the size of the grpah in this case, which will
always be reasonable), it does not set a minimum size.

* Updates the is_valid() base function with more robust checking

* Updates all algorithms such that when they move to an invalid
state they update the interal coloring map to be an empty map. This
should likely be updated to be an uncolored version of the input
map or to have the algorithms return a valid/invalid rather than
the result.

* Retab various files

* Contains enough test cases to give a weak validation to the hybrid
algorithms, and closes #1. It should be noted that these hybrid
algorithms were actually incorrectly coded in the first place and
will need to be majorly refactored through a refactor of Tabucol
which allows for partial colorings to be used.